feat: jitter en piso de followers y columnas de seguidores como INTEGER
- config.py: agrega FOLLOWER_FLOOR_BASE, FOLLOWER_FLOOR_JITTER_LO/HI
  para que el piso de 8 000 varíe con un random entre 0.985 y 1.123
- generar_datos.py: usa los tres parámetros en follower_count();
  convierte INSTAGRAM_FOLLOWER_COUNT y TIKTOK_FOLLOWER_COUNT a Int64
  (nullable integer) para que el CSV salga sin decimales
- cargar_datos.py: cambia esas dos columnas de FLOAT a INTEGER en
  el schema de BigQuery

+++ cargar_datos.py
@@ -24,9 +24,9 @@ SCHEMAS = {
         bigquery.SchemaField('COUNTRY',                   'STRING'),
         bigquery.SchemaField('CATEGORY',                  'STRING'),
         bigquery.SchemaField('INSTAGRAM_HANDLE',          'STRING'),
-        bigquery.SchemaField('INSTAGRAM_FOLLOWER_COUNT',  'FLOAT'),
+        bigquery.SchemaField('INSTAGRAM_FOLLOWER_COUNT',  'INTEGER'),
         bigquery.SchemaField('TIKTOK_HANDLE',             'STRING'),
-        bigquery.SchemaField('TIKTOK_FOLLOWER_COUNT',     'FLOAT'),
+        bigquery.SchemaField('TIKTOK_FOLLOWER_COUNT',     'INTEGER'),
     ],
     'FACTS_REGISTERED_SOCIAL_MEDIA': [
         bigquery.SchemaField('SM_ID',        'INTEGER'),
